fix: split Node-only exports into @omniscribe/shared/node sub-path
Vite/Rollup fails when Node.js-only modules (child_process, fs, util)
are exported from the main shared barrel. Split cli-resolution and
env-utils into @omniscribe/shared/node entry point, add typesVersions
for moduleResolution: node compatibility, and update all plugin imports.
