+++ apps/desktop/src/modules/shared/env-utils.ts
@@ -4,4 +4,4 @@
  * Re-exports from @omniscribe/shared — the canonical source.
  * This file is kept for backward compatibility with existing desktop imports.
  */
-export { ENV_ALLOWLIST, ENV_BLOCKLIST_PATTERNS, buildSafeEnv } from '@omniscribe/shared';
+export { ENV_ALLOWLIST, ENV_BLOCKLIST_PATTERNS, buildSafeEnv } from '@omniscribe/shared/node';
