+++ packages/shared/src/utils/index.ts
@@ -4,15 +4,3 @@ export { extractErrorMessage } from './error';
 export { formatFileSize } from './format';
 export { normalizePath } from './path';
 export { stripAnsiCodes } from './ansi';
-export { joinPaths, getHomeDir, isWindows } from './platform';
-export {
-  findCliCommandSync,
-  findInPathSync,
-  findFirstExistingPath,
-  findCliInPath,
-  findCliInLocalPaths,
-  getNvmBinPaths,
-  getFnmBinPaths,
-  getNvmWindowsCliPaths,
-} from './cli-resolution';
-export { ENV_ALLOWLIST, ENV_BLOCKLIST_PATTERNS, buildSafeEnv } from './env-utils';
